requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
streamlit>=1.28.0
shap>=0.42.0
kafka-python>=2.0.2  